pyproject.toml
[project]
name = "movie-rag-explorer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=0.6.0",
    "sentence-transformers>=3.0.0",
    "streamlit>=1.45.0",
    "requests>=2.32.0",
    "python-dotenv>=1.1.0",
    "ollama>=0.4.0",
]
